requirements.txt
spotipy==2.23.0
python-dateutil==2.9.0
python-dotenv==1.0.1
